refactor: change logo on footer
