+++ index.html
@@ -5,7 +5,7 @@
         <link
             rel="icon"
             type="image/svg+xml"
-            href="./src/assets/images/logo-sementara.png"
+            href="./src/assets/images/logo/logo.png"
         />
         <!-- <meta
             http-equiv="Content-Security-Policy"
