+++ src/types/constant.ts
@@ -21,7 +21,6 @@ export const PAGINATION_PER_PAGE = 5;
 export const PAGE_TITLE = '\uE000nno\uE070nce\uE063ent';
 export const SPONSOR_TITLE = 'S\uE064o\uE01Esore\uE03A B\uE05A:';
 export const MEDPAR_TITLE = '\uE023e\uE053ia Partner';
-export const LAYOUT_TITLE = '\uE028\uE094\uE021\uE086 2022';
 export const REGEX_PARSE = /(\r\n|\n|\r)/gm;
 export const PARTNER_TITLE = '\uE096\uE00Ean\uE01As t\uE032';
 export const ABOUT_US_TITLE = '\uE000\uE050o\uE069\uE067 \uE038\uE076';
